Added 3 HexOrientations, separated functions, removed CalculatePathEvent
