+++ crates/simula_hexgrid/src/lib.rs
@@ -1,2 +1,3 @@
 pub mod hexgrid;
 pub mod pathfinding;
+pub mod user_interface;
